requirements.txt
aiohttp==3.7.4.post0
aiohttp-retry==2.4.5
appdirs==1.4.4
argon2-cffi==21.1.0
async-timeout==3.0.1
atpublic==2.3
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
cachetools==4.2.2
category-encoders==2.2.2
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.6
cloudpickle==2.0.0
colorama==0.4.4
commonmark==0.9.1
configobj==5.0.6
cycler==0.10.0
dask==2021.9.0
debugpy==1.4.3
decorator==5.1.0
defusedxml==0.7.1
dictdiffer==0.9.0
diskcache==5.2.1
distro==1.6.0
dpath==2.0.5
dulwich==0.20.25
dvc==2.7.4
entrypoints==0.3
featurewiz==0.0.42
flatten-dict==0.4.2
flufl.lock==3.2
fsspec==2021.8.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gcsfs==2021.8.1
gitdb==4.0.7
GitPython==3.1.24
google-auth==2.1.0
google-auth-oauthlib==0.4.6
grandalf==0.6
idna==3.2
imbalanced-learn==0.8.0
ipykernel==6.4.1
ipython==7.27.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.1
joblib==1.0.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==7.0.3
jupyter-console==6.4.0
jupyter-core==4.8.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
kiwisolver==1.3.2
locket==0.2.1
mailchecker==4.0.11
MarkupSafe==2.0.1
matplotlib==3.4.3
matplotlib-inline==0.1.3
mistune==0.8.4
multidict==5.1.0
nanotime==0.5.2
nbclient==0.5.4
nbconvert==6.1.0
nbformat==5.1.3
nest-asyncio==1.5.1
networkx==2.6.3
notebook==6.4.4
numpy==1.21.2
oauthlib==3.1.1
packaging==21.0
pandas==1.3.3
pandocfilters==1.5.0
parso==0.8.2
partd==1.2.0
pathspec==0.8.1
patsy==0.5.1
pexpect==4.8.0
phonenumbers==8.12.31
pickleshare==0.7.5
Pillow==8.3.2
ply==3.11
prometheus-client==0.11.0
prompt-toolkit==3.0.20
psutil==5.8.0
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydot==1.4.2
pygit2==1.6.1
Pygments==2.10.0
pygtrie==2.4.2
pyparsing==2.4.7
pyrsistent==0.18.0
python-benedict==0.24.2
python-dateutil==2.8.2
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.1
PyYAML==5.4.1
pyzmq==22.3.0
qtconsole==5.1.1
QtPy==1.11.1
requests==2.26.0
requests-oauthlib==1.3.0
rich==10.10.0
rsa==4.7.2
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
scikit-learn==0.24.2
scipy==1.7.1
seaborn==0.11.2
Send2Trash==1.8.0
shortuuid==1.0.1
shtab==1.4.1
six==1.16.0
smmap==4.0.0
statsmodels==0.12.2
tabulate==0.8.9
terminado==0.12.1
testpath==0.5.0
text-unidecode==1.3
threadpoolctl==2.2.0
toml==0.10.2
toolz==0.11.1
tornado==6.1
tqdm==4.62.2
traitlets==5.1.0
typing-extensions==3.10.0.2
urllib3==1.26.6
voluptuous==0.12.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==3.5.1
xgboost==1.4.2
xlrd==2.0.1
xmltodict==0.12.0
yarl==1.6.3
zc.lockfile==2.0
